+++ string.go
@@ -1,6 +1,8 @@
 package lo
 
-import "unicode/utf8"
+import (
+	"unicode/utf8"
+)
 
 // Substring return part of a string.
 func Substring[T ~string](str T, offset int, length uint) T {
@@ -22,6 +24,36 @@ func Substring[T ~string](str T, offset int, length uint) T {
 	}
 
 	return str[offset : offset+int(length)]
+}
+
+// ChunkString returns an array of strings split into groups the length of size. If array can't be split evenly,
+// the final chunk will be the remaining elements.
+func ChunkString[T ~string](str T, size int) []T {
+	if size <= 0 {
+		panic("lo.ChunkString: Size parameter must be greater than 0")
+	}
+
+	if len(str) == 0 {
+		return []T{""}
+	}
+
+	if size >= len(str) {
+		return []T{str}
+	}
+
+	var chunks []T = make([]T, 0, ((len(str)-1)/size)+1)
+	currentLen := 0
+	currentStart := 0
+	for i := range str {
+		if currentLen == size {
+			chunks = append(chunks, str[currentStart:i])
+			currentLen = 0
+			currentStart = i
+		}
+		currentLen++
+	}
+	chunks = append(chunks, str[currentStart:])
+	return chunks
 }
 
 // RuneLength is an alias to utf8.RuneCountInString which returns the number of runes in string.